requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cors==1.0.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
filelock==3.18.0
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-file==2.1.0
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
zope.event==5.1.1
zope.interface==7.2
